:construction: Create First Component, ServerComponent
This is our first component called ServerComponent. We first need to import Component from @angular/core to use it's @Component decorator. The @Component decorator is a TypeScript feature that enhances your classes and enhances elements in your code. We use the @Component's functionality by identifying the ServerComponents selector name and set up it's HTML template.