:construction: Create 2nd Component, ServersComponent
Run command ng generate component servers. This automatically creates all the files for the new component. Also, automatically adds the new component, ServersComponent to the declarations array in @ngModule.  In the servers.component.html, <app-server> component gets passed and reused 3 times. Finally, in the app.component.html, we pass in the <app-servers> component that holds the 3 <app-server> components